requirements.txt
tensorflow>=2.11.0
numpy>=1.23.0
opencv-python>=4.8.0.76
Pillow>=9.4.0
tk
